split utils.py god-module into classify/qc/marker submodules
- classify_utils.py: relabel, process_query, rfc_pred, classify_cells, aggregate_labels
- qc_utils.py: read_query, is_outlier, qc_preprocess, get_lm, get_qc_metrics
- marker_utils.py: map_celltype_hierarchy, make_stable_colors, get_gene_to_celltype_map, make_celltype_matrices
- dropped unused imports (train_test_split, label_binarize, defaultdict, plt, Path, subprocess, sklearn.metrics *)
- process_markers.py no longer imports from utils (never used any of it)
- added missing explicit 'import random' to process_query.py, process_query_samples.py, scvi_classify.py (previously relied on transitive import via utils.py wildcard import)
- verified end-to-end via fresh pipeline run (PROCESS_QUERY_COMBINED, RF_CLASSIFY, PROCESS_QC, RUN_MULTIQC all pass)
